+++ src/nmt/translate.py
@@ -62,7 +62,7 @@ def add_subparser(subparsers: argparse._SubParsersAction):
                        help='maximum number of tokens')
   group.add_argument('--enc_dropout', default=0.25, type=float,
                       help='encoder dropout rate')
-  group.add_argument('--dec_dropout', default=0.25, type=int,
+  group.add_argument('--dec_dropout', default=0.25, type=float,
                        help='decoder dropout rate')
   
   group = parser.add_argument_group('Vocabulary and model paths')
